+++ eth_defi/erc_4626/core.py
@@ -189,6 +189,23 @@ class ERC4626Feature(enum.Enum):
     #: https://github.com/goatfi/contracts/blob/main/src/infra/multistrategy/Multistrategy.sol
     goat_like = "goat_like"
 
+    #: USD.ai
+    #:
+    #: https://usd.ai/
+    usdai_like = "usdai_like"
+
+    #: AUTO Finance
+    #: Prev. Tokemak Autopilot
+    #:
+    #: https://app.auto.finance/
+    #: https://github.com/Tokemak/v2-core-pub?tab=readme-ov-file
+    autopool_like = "autopool_like"
+
+    #: NashPoint
+    #:
+    #: https://nashpoint.gitbook.io/nashpoint
+    nashpoint_like = "nashpoint_like"
+
 
 def get_vault_protocol_name(features: set[ERC4626Feature]) -> str:
     """Deduct vault protocol name based on Vault smart contract features.
@@ -263,6 +280,12 @@ def get_vault_protocol_name(features: set[ERC4626Feature]) -> str:
         return "Yearn tokenised strategy"
     elif ERC4626Feature.goat_like in features:
         return "Goat Protocol"
+    elif ERC4626Feature.usdai_like in features:
+        return "USDai"
+    elif ERC4626Feature.autopool_like in features:
+        return "AUTO Finance"
+    elif ERC4626Feature.nashpoint_like in features:
+        return "NashPoint"
 
     # No idea
     if ERC4626Feature.erc_7540_like in features: