+++ eth_defi/vault/risk.py
@@ -94,6 +94,9 @@ VAULT_PROTOCOL_RISK_MATRIX = {
     "Yearn v3": VaultTechnicalRisk.lowish,
     "Yearn tokenised strategy": VaultTechnicalRisk.lowish,
     "Goat Protocol": VaultTechnicalRisk.lowish,
+    "USDai": VaultTechnicalRisk.lowish,
+    "AUTO Finance": VaultTechnicalRisk.lowish,
+    "NashPoint": VaultTechnicalRisk.lowish,
 }
 
 #: Particular vaults that are broken, misleading or otherwise problematic.